+++ src/app/app.module.tns.ts
@@ -16,6 +16,7 @@ import { SettingsComponent } from './settings/settings.component';
 import { TaskFormComponent } from './task-form/task-form.component';
 import { TaskListComponent } from './task-list/task-list.component';
 import { ProjectListComponent } from './project-list/project-list.component';
+import { WelcomeComponent } from './welcome/welcome.component';
 
 @NgModule({
     declarations: [
@@ -26,6 +27,7 @@ import { ProjectListComponent } from './project-list/project-list.component';
         TaskFormComponent,
         TaskListComponent,
         ProjectListComponent,
+        WelcomeComponent,
     ],
     entryComponents: [
         NavModalComponent,