+++ src/app/shared/helpers/page.ts
@@ -6,3 +6,7 @@ export function onNavigatedTo(container, handler: () => void) { // eslint-disabl
 export function onNavigatingFrom(container, handler: () => void) { // eslint-disable-line @typescript-eslint/no-unused-vars
     // Do nothing
 }
+
+export function hideActionBar(container) { // eslint-disable-line @typescript-eslint/no-unused-vars
+    // Do nothing
+}