+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.4.0
 
+- Added welcome screen.
 - Use current project filter to prefill project field in task form.
 - Change app icon on iOS.